refactor: decompose peerSync.js god file into shared kernel + per-protocol modules (#1830)
Split the 4,004-line server/services/sharing/peerSync.js (which mixed
subscription CRUD, the outbound push pipeline, the receiver path, media-library
sync, and CoS history/tasks sync) into one concern per file:

- peerSyncShared.js   — state machine, constants, peer-capability helpers, event bus
- peerSyncAssets.js   — asset manifest build/diff/hash + receiver-side asset pull
- peerSyncPush.js     — pushRecordToPeer pipeline + payload builder + schema gate
- peerSyncReceive.js  — applyIncomingPush + reverse-subscribe + merge dispatch
- peerMediaLibrarySync.js — standalone media-library sweep (#1566)
- peerCosSync.js      — CoS history + tasks sync (#1650)

peerSync.js stays the orchestrator (CRUD + debounced push scheduler + recordEvents/
peer:online listeners + registerSubscriptionAdapter wiring) and re-exports the
extracted public surface so every existing importer keeps working unchanged.

peerSyncShared is a leaf with no sibling imports, so the only remaining module
cycle (peerSync <-> peerSyncReceive, call-time-only) is safe under ESM.

Behavior-preserving: the full server suite produces byte-identical pass/fail
results before and after (zero regressions), the DB suite is 156/156 green, and
the server smoke-boots cleanly.
